Improved exeption handling

+++ main.py
@@ -1,10 +1,10 @@
 #!/usr/bin/env python3
 
+from subprocess import TimeoutExpired
 import sys
 sys.path.append("./src/")
 
 from utils import * 
-
 
 def randomName():
     result = ""
@@ -21,28 +21,68 @@ def executeCommand(command, cwdA=None):
     process.wait()
     return process.returncode
 
+def deleteTempFiles(folder, fileName,resultsToo = True):
+    command = "rm -rf " + fileName + ".cpp " + fileName
+    executeCommand(command, folder)
+    if resultsToo:
+        command = "rm -rf results"
+        executeCommand(command, folder)
+
+
 def openJSON(pathName):
-    with open(pathName) as json_file:
-        data = json.load(json_file)
-        return data
+    try:
+        with open(pathName) as json_file:
+            data = json.load(json_file)
+            return data
+    except EnvironmentError: # parent of IOError, OSError *and* WindowsError where available
+        print("Can't open config file!")
+        print("Make sure you choose the correct folder!")
+        exit()
+
 
 def changeSourceFile(folder, fileName):
     randName = randomName()
     
     startTimer = "Timer " + randName + "(\"" + randName + "\");"
     stopTimer  = randName + ".stop();"
+    foundStart = False
+    foundStop  = False
     outputFile = randName + ".cpp"
 
-    with open(folder + outputFile, 'a') as output: 
-        output.write("#include \"../src/Timer.h\"\n")
-        sourceFile = fileinput.input(folder + fileName, inplace=False)
-        for line in sourceFile: 
-            if line.find("//+timer") != -1:
-                output.write(line.rstrip().replace('//+timer', startTimer)+"\n")
-            elif line.find("//-timer") != -1:
-                output.write(line.rstrip().replace('//-timer', stopTimer)+"\n")
-            else:
-                output.write(line.rstrip()+"\n")
+    try:
+        with open(folder + outputFile, 'a') as output: 
+            output.write("#include \"../src/Timer.h\"\n")
+            sourceFile = fileinput.input(folder + fileName, inplace=False)
+            for line in sourceFile: 
+                if line.find("//+timer") != -1 and not foundStart:
+                    output.write(line.rstrip().replace('//+timer', startTimer)+"\n")
+                    foundStart = True
+                elif line.find("//+timer") != -1 and foundStart:
+                    raise ValueError
+                elif line.find("//-timer") != -1 and foundStart and not foundStop:
+                    output.write(line.rstrip().replace('//-timer', stopTimer)+"\n")
+                    foundStop = True
+                elif line.find("//-timer") != -1 and not foundStart:
+                    raise ValueError
+                elif line.find("//-timer") != -1 and foundStop:
+                    raise ValueError
+                else:
+                    output.write(line.rstrip()+"\n")
+            
+            if not foundStart or not foundStop:
+                raise ValueError
+
+    except ValueError:
+        print("The source file is not well formatted")
+        deleteTempFiles(folder, randName)
+        exit()      
+
+    except:
+        print("Could not elaborate the source file!")
+        print("Make sure the file exists!")
+        deleteTempFiles(folder, randName)
+        exit()
+
 
     return randName
 
@@ -62,18 +102,33 @@ def changeCompilerCommand(compilerCommand, sourceFile, outputName):
 
     return compilerCommand
 
-def runBenchmarking(NumberIterations, folder, outputName):
-    with open(folder + "results/data.txt", 'a') as dataFile: 
-        for _ in range(NumberIterations):
-            process = subprocess.Popen("./"+outputName,  shell=True, cwd = folder, stdout=subprocess.PIPE)
-            process.wait()
-            output = process.stdout.read().decode("ascii").split()
-            dataFile.write(output[output.index(outputName)+2] + "\n")
+def runBenchmarking(NumberIterations, folder, outputName, timeoutInput):
+    try:
+        with open(folder + "results/data.txt", 'a') as dataFile: 
+            for _ in range(NumberIterations):
+                process = subprocess.Popen("./"+outputName,  shell=True, cwd = folder, stdout=subprocess.PIPE)
+                process.wait(timeout=timeoutInput)
+                if process.returncode != 0:
+                    raise Exception
 
+                output = process.stdout.read().decode("ascii").split()
+                time = output[output.index(outputName)+2]
+                dataFile.write(time + "\n")
+
+    except TimeoutExpired:
+        print("An execution did not terminated within the timeout!")
+        deleteTempFiles(folder, outputName)
+        exit()        
+
+    except:
+        print("Something went wrong during the benchmarking")
+        deleteTempFiles(folder, outputName)
+        exit()
 
 def main():
 
     if(len(sys.argv)<2):
+        print("No folder name as argument!")
         return
     
     projectName = sys.argv[1]
@@ -82,26 +137,38 @@ def main():
     jsonPath = folder + "config.json"
     config = openJSON(jsonPath)
 
+    neededKeys = [ "ProjectName","SourceFile","CompilerCommand","NumberIterations", "Timeout"]
+    if (neededKeys - config.keys()):
+        print("Config file is not correct. Please check you inserted the right keys!")
+        exit()
+
     sourceFile = config["SourceFile"]
 
     command = "rm -rf results"
-    executeCommand(command, folder)
+    if executeCommand(command, folder)!=0:
+        print("Could not delete results folder!")
+        exit()
     command = "mkdir results"
-    executeCommand(command, folder)
+    if executeCommand(command, folder)!=0:
+        print("Could not create results folder!")
+        exit()
 
     outputName = changeSourceFile(folder, sourceFile)
 
     compilerCommand = changeCompilerCommand(config["CompilerCommand"], sourceFile, outputName)
 
-    executeCommand(compilerCommand, folder)
-
-    runBenchmarking(config["NumberIterations"], folder, outputName)
+    if executeCommand(compilerCommand, folder)!=0:
+        print("Something went wrong during the compiling of code.")
+        print("Make sure you wrote the command correctly")
+        deleteTempFiles(folder, outputName)
+        exit()
+        
+    runBenchmarking(int(config["NumberIterations"]), folder, outputName, config["Timeout"])
 
     analyzer.analyzeData(folder + "results/data.txt")
     plot.plot(config["ProjectName"], folder)
 
-    command = "rm -rf " + outputName + ".cpp " + outputName
-    executeCommand(command, folder)
+    deleteTempFiles(folder, outputName, resultsToo = False)
 
 
 if __name__ == "__main__":
